# specifications/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nEngineering memory for AI product discovery.\n\n**This API has no authentication.** The MVP defers authentication, teams, and\nroles, so it is safe only behind a network boundary (ADR-0014).\n\nLong operations never hold a request open: enqueueing agent work returns `202`\nwith a durable run identifier, and the client polls the run.\n",
    "title": "KAE-Memory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/knowledge/{item_id}/confirm": {
      "post": {
        "description": "Confirm a candidate.\n\nConfirmation is a human act. No agent performs it, which is why this is an\nendpoint and not a step inside a run.\n\n**Transitional (ADR-0007).** This confirms an extracted row. It is not the\nattention queue; ``POST /v1/projects/{project_id}/attention`` is.",
        "operationId": "confirm_knowledge_v1_knowledge__item_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Knowledge",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/projects": {
      "get": {
        "description": "List every project, newest first.",
        "operationId": "list_projects_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "description": "Create a durable project, or return the one that already holds this key.\n\nIdempotent, matching `kae_create_project`. The two surfaces previously\ndisagreed: this endpoint created a second project keyed `name-2` while the\nMCP tool resolved to the existing one. The same request should not mean two\ndifferent things depending on which door it arrived through.\n\n``201`` when a project was created, ``200`` when an existing one was\nreturned, so a caller can tell without comparing identifiers.",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete this project and everything scoped to it, in one transaction.\n\n**Irreversible.** There is no archive and nothing is retained; the\n`deletion-plan` above is the only chance to see what goes.\n\nReturns what was removed rather than an empty body, so a caller can record\nit. One project per call deliberately: a batch is one decision covering many\nprojects, and the ability to delete fifty in a request is the ability to\ndelete forty-nine by accident.",
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "workspace"
        ]
      },
      "get": {
        "description": "Return one project.",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/projects/{project_id}/deletion-plan": {
      "get": {
        "description": "What deleting this project would remove. Changes nothing.\n\nSeparate from the deletion and safe to call repeatedly, because reading the\nlist is the part a person should take their time over. Fifty-five test\nprojects were once cleared by hand-ordered SQL against production; the\ndifference between that and this is that this can be read first.",
        "operationId": "deletion_plan_v1_projects__project_id__deletion_plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deletion Plan",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/projects/{project_id}/knowledge": {
      "get": {
        "description": "List a project's knowledge, and what each statement is about.\n\nUnfiltered by default. The workspace needs to show proposed candidates beside\nconfirmed knowledge, because the difference between them is the thing a user\nacts on.\n\n**Each item now carries its discovery areas.** Memory has always held them\nand this listing did not return them, so a consumer could see what a project\nknows and not what any of it was about. The visible cost was a Definition\npage that reported the problem statement as uncomputable for every project\nin existence \u2014 \"the problem\" is the statements linked to\n`problem_and_value`, and nothing else identifies them.",
        "operationId": "list_knowledge_v1_projects__project_id__knowledge_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeResponse"
                  },
                  "title": "Response List Knowledge V1 Projects  Project Id  Knowledge Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/projects/{project_id}/knowledge/confirm": {
      "post": {
        "description": "Confirm a named set of candidates as one act.\n\n**The reading is what a person agrees to, not the rows underneath it.** A\nconversation synthesises nine statements into one sentence and asks whether\nit holds; the answer is one yes. Confirming per item made a caller decompose\nthat yes itself, and none did \u2014 an interviewer wrote \"Confirmed\" while the\npanel two across read \"0 of 1 confirmed\".\n\nThe set is the turn's provenance: the items the synthesis was drawn from.\nSending it back is what binds agreement to what was actually shown.\n\nAll or nothing, and one revision bump. A partially applied confirmation\nwould leave someone believing they agreed to a reading while part of it\nstayed proposed, and no surface distinguishes that from having agreed.\n\nOn the adapter deliberately. Four capabilities in this codebase were built\ncomplete and reachable from nothing, and the reason readiness reported 0%\nwas a fifth. A service method with passing tests looks healthy from below.",
        "operationId": "confirm_knowledge_set_v1_projects__project_id__knowledge_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmKnowledgeSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeResponse"
                  },
                  "title": "Response Confirm Knowledge Set V1 Projects  Project Id  Knowledge Confirm Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Knowledge Set",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/runs/{run_id}": {
      "get": {
        "description": "Return one run. This is what a client polls while work proceeds.",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "description": "Stream run progress as Server-Sent Events.\n\nA convenience over ``GET /v1/runs/{id}``, never a replacement for it.\nCorrectness must not depend on an uninterrupted browser connection\n(ADR-0009), so a client that misses the stream entirely can still read the\nsame state with an ordinary request.\n\nEmits the current state immediately, then on every change to status,\nattempt, checkpoint, or error. Closes when the run reaches a terminal\nstatus.",
        "operationId": "stream_run_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}